Nomeação, JavaBeans e documentação
Alteração do nome do pacote. Alterações no pacote para servir ao JavaBeans e comentários de documentação adicionados.
