+++ src/br/uniwork/controller/Empresa.java
@@ -1,17 +1,49 @@
-package projeto;
+package br.uniwork.controller;
 /**
+ * Classe que abstrai Conta de empresas no sistema.
  * @author mateus-cabral
  * @version 1.0
  */
 public class Empresa extends Conta{
 	
+	/**
+	 * Nome da empresa
+	 */
 	private String nome;
+	
+	/**
+	 * Segmento de atuação da empresa
+	 */
 	private String segmento;
+	
+	/**
+	 * Cnpj vinculado à empresa
+	 */
 	private String cnpj;
+	
+	/**
+	 * Valor de confiabilidade da empresa
+	 */
 	private float trust;
 	
 	
-	//Construtor
+	/**
+	 * Construtor vazio
+	 */
+	public Empresa() {
+
+	}
+	
+	/**
+	 * Construtor parâmetrizado
+	 * @param nome
+	 * @param id
+	 * @param login
+	 * @param pwd
+	 * @param email
+	 * @param cel
+	 * @param cnpj
+	 */
 	public Empresa(String nome, int id, String login, String pwd, String email, String cel, String cnpj) {
 		super(nome, id, login, pwd, email, cel);
 		this.setCnpj(cnpj);
@@ -32,37 +64,64 @@ public class Empresa extends Conta{
 							"Status Email: " + (this.isStatusEmail() ? "Autenticado" :  "Não autenticado"));
 		
 	}
+
+	//Getters e Setters
 	
+	/**
+	 * @return the nome
+	 */
 	public String getNome() {
 		return nome;
 	}
 
+	/**
+	 * @param nome the nome to set
+	 */
 	public void setNome(String nome) {
 		this.nome = nome;
 	}
 
+	/**
+	 * @return the segmento
+	 */
 	public String getSegmento() {
 		return segmento;
 	}
 
+	/**
+	 * @param segmento the segmento to set
+	 */
 	public void setSegmento(String segmento) {
 		this.segmento = segmento;
 	}
 
-
+	/**
+	 * @return the cnpj
+	 */
 	public String getCnpj() {
 		return cnpj;
 	}
 
+	/**
+	 * @param cnpj the cnpj to set
+	 */
 	public void setCnpj(String cnpj) {
 		this.cnpj = cnpj;
 	}
 
+	/**
+	 * @return the trust
+	 */
 	public float getTrust() {
 		return trust;
 	}
 
+	/**
+	 * @param trust the trust to set
+	 */
 	public void setTrust(float trust) {
 		this.trust = trust;
 	}
+	
+	
 }
