+++ src/br/uniwork/controller/PropostaEmprego.java
@@ -1,22 +1,75 @@
-package projeto;
-
+package br.uniwork.controller;
+/**
+ * Classe que abstrai uma proposta de emprego dentro do sistema.
+ * @author mateus-cabral
+ * @version 1.0
+ */
 import java.util.ArrayList;
 
 public class PropostaEmprego {
 	
+	/**
+	 * Id da proposta no banco de dados
+	 */
 	private int idProposta;
+	
+	/**
+	 * Id da empresa no banco de dados
+	 */
 	private int idEmpresa;
+	
+	/**
+	 * Nome do cargo
+	 */
 	private String nomeCargo;
+	
+	/**
+	 * Descrição do comportamento e responsabilidades esperados
+	 */
 	private String descritivo;
+	
+	/**
+	 * Salário
+	 */
 	private double salario;
+	
+	/**
+	 * Localização do cargo
+	 */
 	private String local;
+	
+	/**
+	 * Lista de habilidades requisitadas
+	 */
 	private ArrayList<Integer> idHabilidade = new ArrayList<Integer>();
+	
+	/**
+	 * Lista de candidatos candidatados
+	 */
 	private ArrayList<Integer> idCandidatos = new ArrayList<Integer>();
+	
+	/**
+	 * Verifica se a proposta está ativa
+	 */
 	private boolean status;
 	
+	/**
+	 * Construtor vazio
+	 */
+	public PropostaEmprego() {
+		
+	}
 	
-	//Construtor parâmetrizado
-	//id habilidades deve receber o id das habilidades necessárias
+	/**
+	 * Construtor parâmetrizado
+	 * @param idProposta int
+	 * @param idEmpresa int
+	 * @param nomeCargo String
+	 * @param descritivo String
+	 * @param salario float
+	 * @param local String
+	 * @param skills ArrayList<Integer>
+	 */
 	public PropostaEmprego(int idProposta, int idEmpresa, String nomeCargo, String descritivo, double salario, String local, ArrayList<Integer> skills) {
 		this.setIdProposta(idProposta);
 		this.setIdEmpresa(idEmpresa);
